@@ -35,11 +35,11 @@
     "normalize-scss": "7.0.1",
     "ramda": "0.27.1",
     "redux": "4.0.4",
-    "redux-vuex": "0.3.2",
+    "redux-vuex": "3.0.0-beta.2",
     "regenerator-runtime": "0.13.3",
     "v-tooltip": "2.0.2",
-    "vue": "2.6.12",
-    "vue-i18n": "8.15.0"
+    "vue": "3.0.4",
+    "vue-i18n": "9.0.0-beta.16"
   },
   "devDependencies": {
     "@podlove/build": "^5.3.1",
