@@ -26,7 +26,7 @@
     "redux-actions": "2.6.5",
     "redux-saga": "1.1.3",
     "redux-vuex": "0.3.2",
-    "vue": "2.6.12",
+    "vue": "3.0.4",
     "vue-i18n": "8.15.0"
   },
   "devDependencies": {
