@@ -20,7 +20,7 @@
     "normalize.css": "8.0.1",
     "rangetouch": "2.0.0",
     "v-tooltip": "2.0.2",
-    "vue": "2.6.12",
+    "vue": "3.0.4",
     "vue-mutation-observer": "0.1.2"
   },
   "devDependencies": {
