@@ -36,8 +36,7 @@ module.exports = {
             rules.style.postcss.plugins.autoprefixer
           ]
         })
-      ]),
-      rules.pug()
+      ])
     ]
   },
 
