@@ -1,5 +1,7 @@
-<template lang="pug">
-  div.episode-subtitle(:style="style" data-test="episode-subtitle") {{ subtitle }}
+<template>
+  <div class="episode-subtitle" :style="style" data-test="episode-subtitle">
+    {{ state.subtitle }}
+  </div>
 </template>
 
 <script>
@@ -7,16 +9,20 @@ import { mapState } from 'redux-vuex'
 import select from 'store/selectors'
 
 export default {
-  data: mapState({
-    font: select.theme.fontRegular,
-    color: select.theme.contrast,
-    subtitle: select.episode.subtitle
-  }),
+  setup() {
+    return {
+      state: mapState({
+        font: select.theme.fontRegular,
+        color: select.theme.contrast,
+        subtitle: select.episode.subtitle
+      })
+    }
+  },
   computed: {
     style() {
       return {
-        color: this.color,
-        ...this.font
+        color: this.state.color,
+        ...this.state.font
       }
     }
   }
