@@ -1,33 +1,50 @@
-<template lang="pug">
-  div.flex.px-2.-mx-2.rounded-sm(:class="{'font-medium': chapter.active}" :style="style" @mouseover="mouseOverHandler" @mouseleave="mouseLeaveHandler" data-test="tab-chapters--entry")
-    span.cursor-pointer.w-8.py-2.mr-2.flex.items-center.justify-center(@click="selectChapter()" aria-hidden="true")
-      icon(v-if="action.icon" :type="action.icon" :size="24" :data-test="`tab-chapters--trigger--${action.icon}`")
-      span(v-else data-test="tab-chapters--index") {{ action.content }}
-
-    chapter-progress.w-full(
+<template>
+  <div
+    class="flex px-2 -mx-2 rounded-sm"
+    data-test="tab-chapters--entry"
+    :class="{ 'font-medium': chapter.active }"
+    :style="style"
+    @mouseover="mouseOverHandler"
+    @mouseleave="mouseLeaveHandler"
+  >
+    <span
+      class="cursor-pointer w-8 py-2 mr-2 flex items-center justify-center"
+      aria-hidden="true"
+      @click="selectChapter()"
+    >
+      <icon
+        v-if="action.icon"
+        :type="action.icon"
+        :size="24"
+        :data-test="`tab-chapters--trigger--${action.icon}`"
+      />
+      <span v-else data-test="tab-chapters--index">{{ action.content }}</span>
+    </span>
+    <chapter-progress
+      class="w-full"
       :chapter="chapter"
-      :showLink="true"
-      :playtime="playtime"
-      :ghost="ghost"
-      :progressColor="progressColor"
+      :show-link="true"
+      :playtime="state.playtime"
+      :ghost="state.ghost"
+      :progress-color="progressColor"
       @chapter="dispatch"
       @play="dispatch"
       @ghost="dispatch"
       @simulate="dispatch"
       @playtime="dispatch"
       @hover="linkMouseover"
-    )
+    />
+  </div>
 </template>
 
 <script>
-import { mapActions } from 'redux-vuex'
+import { mapState, injectStore } from 'redux-vuex'
 import color from 'color'
 import { setChapter } from '@podlove/player-actions/chapters'
 import { requestPlay, requestPause } from '@podlove/player-actions/play'
 import { toHumanTime } from '@podlove/utils/time'
 
 import select from 'store/selectors'
-import store from 'store'
 
 import ChapterProgress from '@podlove/components/chapter-progress'
 import Icon from '@podlove/components/icons'
@@ -45,16 +62,21 @@ export default {
     }
   },
 
-  data() {
+  setup() {
     return {
-      ...this.mapState({
+      state: mapState({
         playtime: select.playtime,
         color: select.theme.brandLightest,
         ghost: select.ghost.time,
         current: select.chapters.current,
         playing: select.driver.playing
       }),
+      dispatch: injectStore().dispatch
+    }
+  },
 
+  data() {
+    return {
       hover: false,
       linkHover: false
     }
@@ -66,7 +88,7 @@ export default {
     },
 
     progressColor() {
-      return color(this.color)
+      return color(this.state.color)
         .alpha(0.5)
         .string()
     },
@@ -74,7 +96,7 @@ export default {
     style() {
       return this.hover
         ? {
-            background: color(this.color)
+            background: color(this.state.color)
               .alpha(0.3)
               .string()
           }
@@ -88,9 +110,9 @@ export default {
         }
       }
 
-      if (this.current.index === this.chapter.index) {
+      if (this.state.current.index === this.chapter.index) {
         return {
-          icon: this.playing ? 'menu-pause' : 'menu-play'
+          icon: this.state.playing ? 'menu-pause' : 'menu-play'
         }
       }
 
@@ -124,18 +146,16 @@ export default {
         return false
       }
 
-      if (this.current.index === this.chapter.index) {
-        store.dispatch(this.playing ? requestPause() : requestPlay())
+      if (this.state.current.index === this.chapter.index) {
+        this.dispatch(this.state.playing ? requestPause() : requestPlay())
         return false
       }
 
-      store.dispatch(setChapter(this.chapter.index - 1))
-      store.dispatch(requestPlay())
+      this.dispatch(setChapter(this.chapter.index - 1))
+      this.dispatch(requestPlay())
       event && event.preventDefault()
       return false
-    },
-
-    dispatch: store.dispatch
+    }
   }
 }
 </script>
