@@ -1,21 +1,53 @@
-<template lang="pug">
-  div.flex.px-2.-mx-2.cursor-pointer.rounded-sm(:class="{'font-medium': episode.active}" :style="style" @mouseover="mouseOverHandler" @mouseleave="mouseLeaveHandler" :data-test="`tab-playlist--entry${episode.active ? '--active' : ''}`")
+<template>
+  <div
+    class="flex px-2 -mx-2 cursor-pointer rounded-sm"
+    :class="{ 'font-medium': episode.active }"
+    :style="style"
+    :data-test="`tab-playlist--entry${episode.active ? '--active' : ''}`"
+    @mouseover="mouseOverHandler"
+    @mouseleave="mouseLeaveHandler"
+  >
+    <span
+      v-if="episode.active"
+      class="w-8 py-2 mr-2 flex justify-center items-center"
+      aria-hidden="true"
+      data-test="tab-playlist--entry--interaction"
+      @click="play()"
+    >
+      <icon :type="playing ? 'menu-pause' : 'menu-play'" :size="24" />
+    </span>
 
-    span.w-8.py-2.mr-2.flex.justify-center.items-center(v-if="episode.active" @click="play()" aria-hidden="true" data-test="tab-playlist--entry--interaction")
-      icon(:type="playing ? 'menu-pause' : 'menu-play'" :size="24")
+    <span
+      v-else
+      class="w-8 py-2 mr-2 flex justify-center items-center"
+      aria-hidden="true"
+      data-test="tab-playlist--entry--interaction"
+      @click="select({ index, play: true })"
+    >
+      <icon v-if="hover" type="menu-play" :size="24" />
+      <icon v-else type="menu-empty" :size="12" />
+    </span>
 
-    span.w-8.py-2.mr-2.flex.justify-center.items-center(v-else @click="select({ index, play: true })" aria-hidden="true" data-test="tab-playlist--entry--interaction")
-      icon(v-if="hover" type="menu-play" :size="24")
-      icon(v-else type="menu-empty" :size="12")
+    <span
+      class="block w-full py-2 mr-2"
+      data-test="tab-playlist--entry--title"
+      @click="select({ index, play: true })"
+      >{{ episode.title }}</span
+    >
 
-    span.block.w-full.py-2.mr-2(@click="select({ index, play: true })" data-test="tab-playlist--entry--title") {{ episode.title }}
-
-    timer.block.w-18.py-2(v-if="episode.duration" :time="episode.duration"  data-test="tab-playlist--entry--timer")
+    <timer
+      v-if="episode.duration"
+      class="block w-18 py-2"
+      :time="episode.duration"
+      data-test="tab-playlist--entry--timer"
+    />
+  </div>
 </template>
 
 <script>
 import color from 'color'
 import { compose } from 'ramda'
+import { mapState, injectStore } from 'redux-vuex'
 import { selectEpisode } from '@podlove/player-actions/playlist'
 import { requestPlay, requestPause } from '@podlove/player-actions/play'
 import Timer from '@podlove/components/timer'
@@ -23,7 +55,6 @@ import Icon from '@podlove/components/icons'
 import { toHumanTime } from '@podlove/utils/time'
 
 import select from 'store/selectors'
-import store from 'store'
 
 export default {
   components: {
@@ -42,12 +73,18 @@ export default {
     }
   },
 
-  data() {
+  setup() {
     return {
-      ...this.mapState({
+      state: mapState({
         color: select.theme.brandLightest,
         playing: select.driver.playing
       }),
+      dispatch: injectStore().dispatch
+    }
+  },
+
+  data() {
+    return {
       hover: false
     }
   },
@@ -59,7 +96,7 @@ export default {
     style() {
       return this.hover
         ? {
-            background: color(this.color)
+            background: color(this.state.color)
               .alpha(0.3)
               .string()
           }
@@ -77,14 +114,16 @@ export default {
     },
 
     play() {
-      if (this.playing) {
-        store.dispatch(requestPause())
+      if (this.state.playing) {
+        this.dispatch(requestPause())
       } else {
-        store.dispatch(requestPlay())
+        this.dispatch(requestPlay())
       }
     },
 
-    select: compose(store.dispatch, selectEpisode)
+    select(position) {
+      this.dispatch(selectEpisode(position))
+    }
   }
 }
 </script>
