@@ -1,10 +1,25 @@
-<template lang="pug">
-  tooltip(trigger="click" :content="$t('MESSAGES.COPIED')" @click="copyLink" :color="color" :background="background")
-    channel(type="link" :color="color" :background="background" :filled="hover" @mouseover.native="hover = true" @mouseleave.native="hover = false")
+<template>
+  <tooltip
+    trigger="click"
+    :content="$t('MESSAGES.COPIED')"
+    :color="state.color"
+    :background="state.background"
+    @click="copyLink"
+  >
+    <channel
+      type="link"
+      :color="state.color"
+      :background="state.background"
+      :filled="hover"
+      @mouseover.native="hover = true"
+      @mouseleave.native="hover = false"
+    />
+  </tooltip>
 </template>
 
 <script>
 import copy from 'copy-to-clipboard'
+import { mapState } from 'redux-vuex'
 import Channel from '@podlove/components/channel'
 import Tooltip from '@podlove/components/tooltip'
 
@@ -15,19 +30,23 @@ export default {
     Channel,
     Tooltip
   },
-  data() {
+  setup() {
     return {
-      hover: false,
-      ...this.mapState({
+      state: mapState({
         link: select.share.link,
         color: select.theme.brandDark,
         background: select.theme.brandLightest
       })
     }
   },
+  data() {
+    return {
+      hover: false
+    }
+  },
   methods: {
     copyLink() {
-      copy(this.link)
+      copy(this.state.link)
     }
   }
 }
