@@ -1,8 +1,17 @@
-<template lang="pug">
-  channel(type="xing" :link="link" :color="color" :background="background" :filled="hover" @mouseover.native="hover = true" @mouseleave.native="hover = false")
+<template>
+  <channel
+    type="xing"
+    :link="state.link"
+    :color="state.color"
+    :background="state.background"
+    :filled="hover"
+    @mouseover.native="hover = true"
+    @mouseleave.native="hover = false"
+  />
 </template>
 
 <script>
+import { mapState } from 'redux-vuex'
 import Channel from '@podlove/components/channel'
 
 import select from 'store/selectors'
@@ -11,17 +20,18 @@ export default {
   components: {
     Channel
   },
-  data() {
+  setup() {
     return {
-      hover: false,
-      ...this.mapState({
-        content: select.share.content,
+      state: this.mapState({
         link: select.share.link,
-        episodeTitle: select.episode.title,
-        playtime: select.playtime,
         color: select.theme.brandDark,
         background: select.theme.alt
       })
+    }
+  },
+  data() {
+    return {
+      hover: false
     }
   }
 }
