@@ -1,28 +1,45 @@
-<template lang="pug">
-  div.w-full(class="mobile:p-4 tablet:p-6" data-test="tab-shownotes")
-    tab-title(@close="closeTab" tab="shownotes") {{ $t('SHOWNOTES.TITLE') }}
+<template>
+  <div class="w-full mobile:p-4 tablet:p-6" data-test="tab-shownotes">
+    <tab-title tab="shownotes" @close="closeTab">
+      {{ $t('SHOWNOTES.TITLE') }}
+    </tab-title>
 
-    h4.text-sm.font-bold {{ showTitle }}
-    h3.text-lg.font-bold.mb-4 {{ episodeTitle }}
+    <h4 class="text-sm font-bold">
+      {{ state.showTitle }}
+    </h4>
+    <h3 class="text-lg font-bold mb-4">
+      {{ state.episodeTitle }}
+    </h3>
 
-    p.text-base.font-semibold.mb-3 {{ subtitle }}
+    <p class="text-base font-semibold mb-3">
+      {{ state.subtitle }}
+    </p>
 
-    p.mb-3(v-html="summary")
+    <p class="mb-3" v-html="state.summary"></p>
 
-    divider.mb-4
+    <divider class="mb-4" />
 
-    div(v-for="(group, gid) in groups" :key="`group-${gid}`")
-      h3.text-sm.font-bold.mb-4 {{ group.title }}
-      div.flex.flex-wrap
-        div(class="flex items-center mr-16 last:mr-0 mb-4" v-for="(contributor, cid) in group.contributors" :key="`group-${gid}-contributor-${cid}`")
-          img.rounded-full.h-16(:src="contributor.avatar")
-          span.text-base.font-semibold.pr-4.pl-4 {{ contributor.name }}
+    <div v-for="(group, gid) in state.groups" :key="`group-${gid}`">
+      <h3 class="text-sm font-bold mb-4">
+        {{ group.title }}
+      </h3>
+      <div class="flex flex-wrap">
+        <div
+          v-for="(contributor, cid) in group.contributors"
+          :key="`group-${gid}-contributor-${cid}`"
+          class="flex items-center mr-16 last:mr-0 mb-4"
+        >
+          <img class="rounded-full h-16" :src="contributor.avatar" />
+          <span class="text-base font-semibold pr-4 pl-4">{{ contributor.name }}</span>
+        </div>
+      </div>
+    </div>
+  </div>
 </template>
 
 <script>
-import { mapState } from 'redux-vuex'
+import { mapState, injectStore } from 'redux-vuex'
 import { toggleTab } from '@podlove/player-actions/tabs'
-import store from 'store'
 import select from 'store/selectors'
 
 import TabTitle from '../tab-title'
@@ -34,17 +51,22 @@ export default {
     Divider
   },
 
-  data: mapState({
-    showTitle: select.show.title,
-    episodeTitle: select.episode.title,
-    subtitle: select.episode.subtitle,
-    summary: select.episode.summary,
-    groups: select.contributors.groups
-  }),
+  setup() {
+    return {
+      state: mapState({
+        showTitle: select.show.title,
+        episodeTitle: select.episode.title,
+        subtitle: select.episode.subtitle,
+        summary: select.episode.summary,
+        groups: select.contributors.groups
+      }),
+      dispatch: injectStore().dispatch
+    }
+  },
 
   methods: {
     closeTab() {
-      store.dispatch(toggleTab('shownotes'))
+      this.dispatch(toggleTab('shownotes'))
     }
   }
 }
