@@ -1,17 +1,19 @@
-<template lang="pug">
-  div(data-test="tab-transcripts")
-    div.header(class="mobile:p-4 tablet:p-6")
-      tab-title(@close="closeTab" tab="transcripts") {{ $t('TRANSCRIPTS.TITLE') }}
-      search.mb-6
-    //- Render
-    render-container(:prerender="prerender" v-if="prerender.length > 0")
-    //- Prerender
-    prerender-container(:transcripts="timeline" @load="loadPrerender" v-else)
+<template>
+  <div data-test="tab-transcripts">
+    <div class="header mobile:p-4 tablet:p-6">
+      <tab-title tab="transcripts" @close="closeTab">
+        {{ $t('TRANSCRIPTS.TITLE') }}
+      </tab-title>
+      <search class="mb-6" />
+    </div>
+    <render-container v-if="prerender.length > 0" :prerender="prerender" />
+    <prerender-container v-else :transcripts="state.timeline" @load="loadPrerender" />
+  </div>
 </template>
 
 <script>
+import { mapState, injectStore } from 'redux-vuex'
 import { toggleTab } from '@podlove/player-actions/tabs'
-import store from 'store'
 import select from 'store/selectors'
 
 import TabTitle from '../tab-title'
@@ -27,15 +29,16 @@ export default {
     RenderContainer,
     PrerenderContainer
   },
-  data() {
+  setup() {
     return {
-      ...this.mapState({
-        activeTranscript: select.transcripts.active,
-        activeFollow: select.transcripts.follow,
-        selectedSearch: select.transcripts.searchSelected,
-        searchQuery: select.transcripts.searchQuery,
+      state: mapState({
         timeline: select.transcripts.timeline
       }),
+      dispatch: injectStore().dispatch
+    }
+  },
+  data() {
+    return {
       prerender: []
     }
   },
@@ -54,7 +57,7 @@ export default {
       this.prerender = []
     },
     closeTab() {
-      store.dispatch(toggleTab('transcripts'))
+      this.dispatch(toggleTab('transcripts'))
     }
   }
 }
