@@ -1,18 +1,19 @@
-<template lang="pug">
-  div
-    transcript-entry(
+<template>
+  <div>
+    <transcript-entry
       v-for="(entry, index) in transcripts"
       :key="index"
       :entry="entry"
       :prerender="true"
-      :playtime="playtime"
-      :search-query="searchQuery"
-      :ghost-active="ghostActive"
-      :ghost-time="ghostTime"
-      :search-results="searchResults"
+      :playtime="state.playtime"
+      :search-query="state.searchQuery"
+      :ghost-active="state.ghostActive"
+      :ghost-time="state.ghostTime"
+      :search-results="state.searchResults"
       :chapter-style="chapterStyle"
       :speaker-style="speakerStyle"
-    )
+    />
+  </div>
 </template>
 
 <script>
@@ -33,24 +34,28 @@ export default {
       default: () => []
     }
   },
-  data: mapState({
-    playtime: select.playtime,
-    ghostActive: select.ghost.active,
-    ghostTime: select.ghost.time,
-    searchResults: select.transcripts.searchResults,
-    searchQuery: select.transcripts.searchQuery,
-    fontCi: select.theme.fontCi,
-    fontBold: select.theme.fontBold
-  }),
+  setup() {
+    return {
+      state: mapState({
+        playtime: select.playtime,
+        ghostActive: select.ghost.active,
+        ghostTime: select.ghost.time,
+        searchResults: select.transcripts.searchResults,
+        searchQuery: select.transcripts.searchQuery,
+        fontCi: select.theme.fontCi,
+        fontBold: select.theme.fontBold
+      })
+    }
+  },
   computed: {
     chapterStyle() {
       return {
-        ...this.fontCi
+        ...this.state.fontCi
       }
     },
     speakerStyle() {
       return {
-        ...this.fontBold
+        ...this.state.fontBold
       }
     }
   },
