@@ -1,5 +1,13 @@
-<template lang="pug">
-  timer(role="timer" :color="color" :time="duration - (ghost ? ghost : playtime)" :aria-label="$t(a11y.key, a11y.attr)" tabindex="0" :remaining="true" data-test="timer-duration")
+<template>
+  <timer
+    role="timer"
+    :color="state.color"
+    :time="state.duration - (state.ghost ? state.ghost : state.playtime)"
+    :aria-label="$t(state.a11y.key, state.a11y.attr)"
+    tabindex="0"
+    :remaining="true"
+    data-test="timer-duration"
+  />
 </template>
 
 <script>
@@ -12,12 +20,16 @@ import select from 'store/selectors'
 export default {
   components: { Timer },
 
-  data: mapState({
-    playtime: select.playtime,
-    duration: select.duration,
-    ghost: select.ghost.time,
-    color: select.theme.contrast,
-    a11y: select.accessibility.timerDuration
-  })
+  setup() {
+    return {
+      state: mapState({
+        playtime: select.playtime,
+        duration: select.duration,
+        ghost: select.ghost.time,
+        color: select.theme.contrast,
+        a11y: select.accessibility.timerDuration
+      })
+    }
+  }
 }
 </script>
